requirements.txt
oauth2client
selenium
pillow
pyocr
gspread
python-dotenv